requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
redis==3.3.11
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
